gh-pages branch for the device check
Create a default page for the gh-pages branch on this repo to show the
device check page with some updates.
The new device check page is designed to capture the device information
from the users browser and submit it as a github issue to the
repository.
Future versions could probably run a quick client side check to see if
the device already exists.